requirements.txt
asgiref==3.8.1
astroid==3.3.9
attrs==25.4.0
black==25.1.0
click==8.1.8
dill==0.3.9
Django==5.1.7
djangorestframework==3.15.2
drf-spectacular==0.29.0
flake8==7.2.0
gunicorn==25.0.3
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.11
pycodestyle==2.13.0
pyflakes==3.3.2
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.5
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
tomlkit==0.13.2
uritemplate==4.2.0
uuid==1.30
whitenoise==6.11.0
